requirements.txt
huawei-lte-api
python-dotenv
speedtest-cli
